requirements.txt
requests
beautifulsoup4
pandas
pyyaml
python-dotenv
google-generativeai
